+++ src/dsp/voice.rs
@@ -62,10 +62,10 @@ pub struct Voice {
     pub noise_on: bool,
     pub echo_on: bool,
 
-    sample_start_address: u32,
-    loop_start_address: u32,
+    sample_start_address: u16,
+    loop_start_address: u16,
     brr_block_decoder: BrrBlockDecoder,
-    sample_address: u32,
+    sample_address: u16,
     sample_pos: i32,
 
     pub resampling_mode: ResamplingMode,
@@ -224,7 +224,7 @@ impl Voice {
     }
 
     fn read_entry(&mut self) {
-        let source = self.source as i32;
+        let source = self.source as u16;
         self.sample_start_address = self.dsp().read_source_dir_start_address(source);
         self.loop_start_address = self.dsp().read_source_dir_loop_address(source);
     }
@@ -233,7 +233,7 @@ impl Voice {
         let mut buf = [0; 9];
         let sample_address = self.sample_address;
         for i in 0..9 {
-            buf[i] = self.emulator().read_u8(sample_address + (i as u32));
+            buf[i] = self.emulator().read_u8(sample_address + i as u16);
         }
         self.brr_block_decoder.read(&buf);
         self.sample_address += 9;
